organise file for kemono (wip)
